backend/requirements.txt
# FastAPI + Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
websockets==14.1

# Database - PostgreSQL
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Database - Neo4j
neo4j==5.27.0

# Database - Qdrant
qdrant-client==1.12.1

# Redis + Celery
redis==5.2.1
celery[redis]==5.4.0

# Firebase Admin
firebase-admin==6.6.0

# Scraping
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.27.1
webdriver-manager==4.0.2
newspaper3k==0.2.8
scrapling==0.2.99
crawl4ai==0.4.248
requests==2.32.3
httpx==0.27.2
aiohttp==3.11.11

# NLP + ML
spacy==3.8.4
transformers==4.47.1
torch==2.5.1
safetensors==0.4.5
sentence-transformers==3.3.1

# Telegram Bot
python-telegram-bot==21.9

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
pypdf==5.1.0
python-docx==1.1.2
